requirements.txt
requests==2.28.2
environs==9.5.0
python-telegram-bot==11.1.0
redis==3.2.1